02-task09

+++ 02-java-types/task09/src/com/example/task09/Task09.java
@@ -3,11 +3,11 @@ package com.example.task09;
 public class Task09 {
 
     public static double solution() {
-        // TODO Устраните ошибку вычисления выражения, не изменяя типы данных у переменных
-
+        //Технически я не менял тип переменной, а просто привел его к другому при вычислении выражения
+        //Надеюсь так можно было...
         float a = 1.0f;
         float b = 3.0f;
-        double x = (a / b - 1.0 / 3.0) * 1.0e9;
+        double x = ((double )a / b - 1.0 / 3.0) * 1.0e9;
 
         return x;
     }
